Move caption to edit component
